pyproject.toml
[project]
name = "breathing-rate"
version = "0.1.0"
description = "Compute breathing rate at 15-second intervals from RR-interval logs."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "python-dateutil>=2.8",
]

[project.scripts]
breathing-rate = "src.compute_breathing_rate:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
